Added Docs blocks and @SurrealDbOnlyApi to 'kill'
