+++ src/commonMain/kotlin/uk/gibby/driver/rpc/Live.kt
@@ -5,24 +5,19 @@ import kotlinx.serialization.json.buildJsonArray
 import kotlinx.serialization.json.decodeFromJsonElement
 import uk.gibby.driver.Surreal
 import uk.gibby.driver.annotation.SurrealDbNightlyOnlyApi
-import uk.gibby.driver.model.Thing
 import uk.gibby.driver.surrealJson
 
 
+/**
+ * Live
+ *
+ * This method will create a live query
+ *
+ * @param table Name of the table to 'LIVE SELECT' from
+ * @return The id of the live query
+ */
 @SurrealDbNightlyOnlyApi
 suspend fun Surreal.live(table: String): String {
     val response = sendRequest("live", buildJsonArray { add(table) })
-    return surrealJson.decodeFromJsonElement(response)
-}
-
-@SurrealDbNightlyOnlyApi
-suspend fun Surreal.live(table: String, id: String): String {
-    val response = sendRequest("live", buildJsonArray { add("$table:$id") })
-    return surrealJson.decodeFromJsonElement(response)
-}
-
-@SurrealDbNightlyOnlyApi
-suspend fun Surreal.live(id: Thing<*>): String {
-    val response = sendRequest("live", buildJsonArray { add(id.id) })
     return surrealJson.decodeFromJsonElement(response)
 }
